Extract startup action
